refactor: replace helpers functions into special folder, use default func of formik for reset
